feat: refactor SkipToMain and SkipToToolsMenu components to use SkipLink for improved code reuse
