+++ packages/pxweb2/src/app/components/SkipToMain/SkipToToolsMenu.tsx
@@ -1,48 +1,19 @@
 import React from 'react';
-import cl from 'clsx';
-import { useTranslation } from 'react-i18next';
-import { useLocation, useSearchParams } from 'react-router';
 
-import classes from './SkipToMain.module.scss';
-import { Link } from '@pxweb2/pxweb2-ui';
-import { getConfig } from '../../util/config/getConfig';
-import { getLanguagePath } from '../../util/language/getLanguagePath';
+import { SkipLink } from './SkipLink';
 
 export const SkipToToolsMenu = React.forwardRef<
-  HTMLDivElement,
-  React.HTMLAttributes<HTMLDivElement>
+	HTMLDivElement,
+	React.HTMLAttributes<HTMLDivElement>
 >((props, ref) => {
-  const { t, i18n } = useTranslation();
-  const config = getConfig();
-  const location = useLocation().pathname;
-  const [searchParams] = useSearchParams();
-
-  const basePath = getLanguagePath(
-    location,
-    i18n.language,
-    config.language.supportedLanguages,
-    config.language.defaultLanguage,
-    config.language.showDefaultLanguageInPath,
-    config.baseApplicationPath,
-    config.language.positionInPath,
-  );
-
-  // build a single, URL-encoded query string from the search params
-  const paramsString = searchParams.toString(); // URLSearchParams handles encoding
-  const params = paramsString ? `?${paramsString}` : '';
-  const path = `${basePath}${params}#px-table-viewer-outer-container`;
-
-  return (
-    <div
-      ref={ref}
-      className={cl(classes['skip-to-main'], classes['screen-reader-only'])}
-      {...props}
-    >
-      <Link href={path} size="medium">
-        {t('common.skip_to_toolsmenu')}
-      </Link>
-    </div>
-  );
+	return (
+		<SkipLink
+			ref={ref}
+			targetId="px-table-viewer-outer-container"
+			translationKey="common.skip_to_toolsmenu"
+			{...props}
+		/>
+	);
 });
 
 SkipToToolsMenu.displayName = 'SkipToToolsMenu';
